requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
pydantic[email]>=2.7.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.3
openai>=1.35.0
python-multipart>=0.0.9
